Restucture directories and files of blog metric chart builders
